1687: Problem assigning newly created metadata to groups on IE 1796: IE10: intern of extern: ander uitzicht 1856: IE10 issues - beheer menu icoontjes 1857: IE10 - lijst van groepen is leeg 1858: IE10 - validatierapporten komen er niet door 2001: IE10: content is not allowed in prolog bij editeren van metadata record 2038: IE10 compatibiliteit 2065: Informatie in mails aanpassen 2066: MAil naar administrators bij goedkeuring 2175: Gedrag opgeladen vbweergave verschilt van vroeger opgeladen vbweergaves 2212: Witruimte in mails 2221: Dubbele mails worden verstuurd - capitalisatie issue 2246: operatesOn - label van 2de item is verkeerd 2247: De rechten van templates worden bepaald op basis van reeds aanwezige groepen 2256: Na klikken op thumbnial wordt grote versie niet altijd getoond in nieuw venster

+++ jeeves/src/main/java/jeeves/server/context/ServiceContext.java
@@ -75,7 +75,7 @@ public class ServiceContext extends BasicContext
 	private String service;
 	private String ipAddress;
 	private String uploadDir;
-	private int    maxUploadSize;
+	private double    maxUploadSize;
 	private JeevesServlet servlet;
 	private boolean startupError = false;
 	Map<String,String> startupErrors;
@@ -104,7 +104,7 @@ public class ServiceContext extends BasicContext
 	public String getService()   { return service;   }
 	public String getIpAddress() { return ipAddress; }
 	public String getUploadDir() { return uploadDir; }
-    public int getMaxUploadSize() { return maxUploadSize; }
+    public double getMaxUploadSize() { return maxUploadSize; }
 
 	public UserSession    getUserSession()    { return userSession; }
 	public ProfileManager getProfileManager() { return profilMan;   }
@@ -124,7 +124,7 @@ public class ServiceContext extends BasicContext
 	public void setServlet (JeevesServlet serv)    { servlet  = serv;    }
 	public void setIpAddress(String address) { ipAddress = address; }
 	public void setUploadDir(String dir)     { uploadDir = dir;     }
-    public void setMaxUploadSize(int size)   { maxUploadSize = size;}
+    public void setMaxUploadSize(double size)   { maxUploadSize = size;}
     public void setStartupErrors(Map<String,String> errs)   { startupErrors = errs; startupError = true; }
 
 	public void setInputMethod (InputMethod m)  { input  = m; }
